requirements.txt
streamlit>=1.30.0
requests>=2.31.0
numpy>=1.25.0
sentence-transformers>=2.2.2
groq>=0.4.0
python-dotenv>=1.0.0

